+++ registry/entries/packages/use-shortcut.ts
@@ -1,0 +1,6 @@
+import sourceConfig from "../../../apps/docs/src/config/packages/use-shortcut";
+import type { PackageConfig } from "../../types";
+
+const useShortcutEntry: PackageConfig = sourceConfig;
+
+export default useShortcutEntry;